# backend_node/src/docs/swagger.json
{
	"openapi": "3.0.0",
	"components": {
		"examples": {},
		"headers": {},
		"parameters": {},
		"requestBodies": {},
		"responses": {},
		"schemas": {
			"Skill": {
				"properties": {
					"id": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"developers": {
						"items": {
							"$ref": "#/components/schemas/Developer"
						},
						"type": "array"
					},
					"tasks": {
						"items": {
							"$ref": "#/components/schemas/Task"
						},
						"type": "array"
					}
				},
				"required": [
					"id",
					"name",
					"developers",
					"tasks"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Task": {
				"properties": {
					"id": {
						"type": "string"
					},
					"title": {
						"type": "string"
					},
					"status": {
						"type": "string"
					},
					"requiredSkills": {
						"items": {
							"$ref": "#/components/schemas/Skill"
						},
						"type": "array"
					},
					"assignedDeveloper": {
						"allOf": [
							{
								"$ref": "#/components/schemas/Developer"
							}
						],
						"nullable": true
					},
					"subtasks": {
						"items": {
							"$ref": "#/components/schemas/Task"
						},
						"type": "array"
					},
					"parentTask": {
						"allOf": [
							{
								"$ref": "#/components/schemas/Task"
							}
						],
						"nullable": true
					}
				},
				"required": [
					"id",
					"title",
					"status",
					"requiredSkills",
					"assignedDeveloper",
					"parentTask"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Developer": {
				"properties": {
					"id": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"skills": {
						"items": {
							"$ref": "#/components/schemas/Skill"
						},
						"type": "array"
					},
					"tasks": {
						"items": {
							"$ref": "#/components/schemas/Task"
						},
						"type": "array"
					}
				},
				"required": [
					"id",
					"name",
					"skills",
					"tasks"
				],
				"type": "object",
				"additionalProperties": false
			},
			"CreateTaskRequest": {
				"properties": {
					"title": {
						"type": "string"
					},
					"requiredSkillIds": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"subtasks": {
						"items": {
							"$ref": "#/components/schemas/CreateTaskRequest"
						},
						"type": "array"
					},
					"parentTaskId": {
						"type": "string"
					}
				},
				"required": [
					"title"
				],
				"type": "object",
				"additionalProperties": false
			}
		},
		"securitySchemes": {
			"bearerAuth": {
				"type": "http",
				"scheme": "bearer",
				"bearerFormat": "JWT"
			}
		}
	},
	"info": {
		"title": "backend_node",
		"version": "1.0.0",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"paths": {
		"/tasks": {
			"get": {
				"operationId": "GetTasks",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/Task"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Tasks"
				],
				"security": [],
				"parameters": []
			},
			"post": {
				"operationId": "CreateTask",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Task"
								}
							}
						}
					}
				},
				"tags": [
					"Tasks"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateTaskRequest"
							}
						}
					}
				}
			}
		},
		"/tasks/assign": {
			"post": {
				"operationId": "AssignTask",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"message": {
											"type": "string"
										}
									},
									"required": [
										"message"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"Tasks"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"developerId": {
										"type": "string"
									},
									"taskId": {
										"type": "string"
									}
								},
								"required": [
									"developerId",
									"taskId"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/tasks/{taskId}/unassign": {
			"post": {
				"operationId": "UnassignTask",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"message": {
											"type": "string"
										}
									},
									"required": [
										"message"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"Tasks"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "taskId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/tasks/{taskId}/status": {
			"patch": {
				"operationId": "UpdateTaskStatus",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"message": {
											"type": "string"
										}
									},
									"required": [
										"message"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"Tasks"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "taskId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"status": {
										"type": "string"
									}
								},
								"required": [
									"status"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/skills": {
			"get": {
				"operationId": "GetSkills",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Skills"
				],
				"security": [],
				"parameters": []
			}
		},
		"/developers": {
			"get": {
				"operationId": "GetDevelopers",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Developers"
				],
				"security": [],
				"parameters": []
			}
		}
	},
	"servers": [
		{
			"url": "/api"
		}
	]
}